+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
+### Changed
+
+- `exec` command now looks for the first session with address instead of failing
+
 ## [2.0.5] - 2024-05-06
 
 ### Fixed